+++ application/tests/models/Sites_Model_test.php
@@ -43,6 +43,13 @@ class Site_Model_test extends TestCase {
 		$this->assertRegExp('/^[c|v][0-9\.]+(?:\/c[0-9\.]+)?$/', $result['latest_chapter']);
 		$this->assertRegExp('/^[0-9]+-[0-9]+-[0-9]+ [0-9]+:[0-9]+:[0-9]+$/', $result['last_updated']);
 	}
+	public function test_MangaHere_fail() {
+		MonkeyPatch::patchFunction('log_message', NULL, 'MangaHere'); //Stop logging stuff...
+		$result = $this->Sites_Model->{'MangaHere'}->getTitleData('i_am_a_bad_url');
+
+		$this->assertNull($result);
+		MonkeyPatch::verifyInvokedOnce('log_message', ['error', 'MangaHere : i_am_a_bad_url | Failure string matched']);
+	}
 
 	public function test_Batoto() {
 		$this->skipTravis('Missing required cookies.');